+++ app.js
@@ -4,7 +4,7 @@ const express = require("express");
 const mongoose = require("mongoose");
 const morgan = require("morgan");
 const https = require("https");
-const { findExpiringListings } = require("./helpers/reminder.helper");
+const { findExpiringListings } = require("./helpers/expiredListings.helper");
 
 const { DATABASE_URI, environment } = require("./config");
 const loginRouter = require("./routes/login");
